Restrict access to CRUD actions for formats, students, teachers, courses controllers by using the logged in clinic. Register movements for accounts, courses and formats

+++ app/Http/Controllers/CoursesController.php
@@ -11,7 +11,9 @@ use App\Teacher;
 use App\Clinic;
 use App\Account;
 use App\Role;
+use App\Movement;
 use App\Http\Requests;
+use Illuminate\Pagination\LengthAwarePaginator;
 
 class CoursesController extends Controller
 {
@@ -19,7 +21,7 @@ class CoursesController extends Controller
     {
         $this->middleware('auth');
         $this->middleware('privileges:courses');
-        $this->middleware('course', ['except' => ['index', 'create', 'store']]);
+        $this->middleware('check.clinic:course', ['except' => ['index', 'create', 'store']]);
     }
     
     /**
@@ -27,10 +29,21 @@ class CoursesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $courses = Course::fromClinic(clinic())->with('group')->with('period')->with('subject')->Paginate(10);
-        return View('courses.index')->with('courses',$courses);
+        // Get current page form url e.g. &page=6
+        $currentPage = LengthAwarePaginator::resolveCurrentPage();
+        // Define how many items we want to be visible in each page
+        $perPage = 20;
+
+        $courses = Course::fromClinic(clinic())->orderBy('period_id', 'DESC')->orderBy('subject_id', 'DESC')->with('group')->with('period')->with('subject')->withCount('students')->skip(($currentPage - 1) * $perPage)->limit($perPage)->get();
+
+        // Create our paginator and pass it to the view
+        $courses = new LengthAwarePaginator($courses, Course::count(), $perPage);
+        $courses->setPath($request->url());
+        $courses->appends($request->except(['page']));
+
+        return View('courses.index')->with('courses', $courses);
     }
 
     /**
@@ -55,15 +68,20 @@ class CoursesController extends Controller
      */
     public function store(Request $request)
     {
-        try{
-            $this->makeValidation($request);
-
-        } catch(\Illuminate\Database\QueryException $e){
-            session()->flash('warning', 'No se pudo crear el grupo error:'.$e->getCode());
-            return redirect('courses/create');
+        if($this->valid_data($request)) {
+            if(!Course::fromClinic(clinic())->withData($request)->exists()) {
+                $this->saveResource($request);
+                Movement::register(account(), null, 'courses.store'); // update course
+                session()->flash('success', 'El curso se creo correctamente');
+                return redirect('courses');
+            } else {
+                session()->flash('warning', 'Ya existe un curso con esos datos');
+            }
+        } else {
+            session()->flash('warning', 'No se pudo crear el curso');
         }
-        session()->flash('success', 'El curso fue creado correctamente.');
-        return redirect('courses');
+
+        return redirect()->back()->withInput($request->all());
     }
 
     /**
@@ -97,7 +115,7 @@ class CoursesController extends Controller
         $groups = Group::pluck('group_id', 'group_id');
         $periods = Period::pluck('period_id', 'period_id');
         $subjects = Subject::pluck('subject_name','subject_id');
-        // $course = Course::fromClinic(clinic())->find($course);
+
         return View('courses.edit')->with('course', $course)->with('groups', $groups)->with('periods', $periods)->with('subjects', $subjects);
     }
 
@@ -110,14 +128,20 @@ class CoursesController extends Controller
      */
     public function update(Request $request, Course $course)
     {
-        try{
-            $this->makeValidation($request, $course);
-        } catch(\Illuminate\Database\QueryException $e){
-            session()->flash('warning', 'El curso no se pudo modificar');
-            return redirect('courses');
+        if($this->valid_data($request)) {
+            if(!Course::fromClinic(clinic())->withData($request)->where('course_id', '<>', $course->course_id)->exists()) {
+                $this->saveResource($request, $course);
+                Movement::register(account(), null, 'courses.update'); // update course
+                session()->flash('success', 'Se modifico el curso exitosamente');
+                return redirect('courses');
+            } else {
+                session()->flash('warning', 'Ya existe un curso con esos datos');
+            }
+        } else {
+            session()->flash('warning', 'No se pudo crear el curso');
         }
-        session()->flash('info', 'Se modifico el curso: '.$course->course_id);
-        return redirect('courses');
+
+        return redirect()->back();
     }
 
     /**
@@ -130,11 +154,12 @@ class CoursesController extends Controller
     {
         try {
             $course->delete();
+            Movement::register(account(), null, 'courses.destroy'); // destroy course
+            session()->flash('success', 'El curso fue eliminado correctamente.');
         } catch (\Illuminate\Database\QueryException $e){
             session()->flash('warning', 'El curso no se puede eliminar');
             return redirect('courses');
         }
-        session()->flash('danger', 'El curso fue eliminado correctamente.');
         return redirect('courses');        
     }
 
@@ -144,14 +169,16 @@ class CoursesController extends Controller
 
         if(isset($account)) {
             if(!$account->has_role('teacher')) {
-                $account->roles()->attach(Role::name('teacher')->role_id); // translate later
+                $account->roles()->attach(config('id.teacher')); // translate later
             }
             if(!$account->has_profile('teacher')) {
-                $account->assign_type('teacher');
+                $account->assign_profile('teacher');
             }
-
-            $course->teachers()->attach($account->user_id);
-            session()->flash('success', 'El profesor fue asignado correctamente.');
+            if(!$course->have_teacher($account)) {
+                $course->teachers()->attach($account->user_id);
+                Movement::register(account(), $account, 'courses.store_teacher'); // store teacher
+                session()->flash('success', 'El profesor fue asignado correctamente.');
+            }
         } else {
             session()->flash('danger', 'Hubo un error al asignar al profesor.');
         }
@@ -160,29 +187,40 @@ class CoursesController extends Controller
 
     public function destroy_teacher(Course $course, Teacher $teacher)
     {
-        $course->teachers()->detach($teacher->user_id);
-        session()->flash('success', 'El profesor fue eliminado correctamente.');
+        if($course->have_teacher($teacher)) {
+            $course->teachers()->detach($teacher->user_id);
+            Movement::register(account(), $teacher->account(clinic()), 'courses.destroy_teacher'); // store teacher
+            session()->flash('success', 'El profesor fue eliminado correctamente.');
+        }
         return redirect()->back(); 
     }
 
-    private function makeValidation(Request $request, $resource = null) 
-    {   
-        $group = Group::findOrFail($request->group_id);
-        $subject = Subject::findOrFail($request->subject_id);
-        $period = Period::findOrFail($request->period_id);
+    private function valid_data($request) {
+        $group = Group::find($request->group_id);
+        $subject = Subject::find($request->subject_id);
+        $period = Period::find($request->period_id);
 
+        if(isset($group) && isset($subject) && isset($period)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    private function saveResource(Request $request, $resource = null) 
+    {   
         $values = ['subject_id' => $request->subject_id, 
                    'group_id' => $request->group_id, 
                    'period_id' => $request->period_id];
 
         if(isset($resource)) {
             $resource->update($values);
-            $resource->generatePk();
+            $resource->generatePK();
             return $resource->update($values);
         } else {
             $resource = new Course($values);
             $resource->generatePK();
-            $resource->clinic_id = Clinic::first()->clinic_id;
+            $resource->clinic_id = clinic()->clinic_id;
             return $resource->save();
         }
     }
